requirements.txt
django==5.2.3
django-extensions==4.1
djangorestframework==3.16.0